Fix some warnings on GCC13/Ubuntu 24.04 (#4265)
* nav_2d_utils: use get_parameter without default value

the original prototype construct a Parameter from the &parameter
references. Our value is unassigned, which caused a warning.
fix ‘value’ may be used uninitialized warning on gcc13.

Signed-off-by: Chuanhong Guo <gch981213@gmail.com>

* nav2_behaviors: use get_parameter without default value

the two-parameter version of get_parameter constructs a
Parameter from the &parameter references. Our value is
unassigned, which caused a warning.
fix 'may be used uninitialized' warning on gcc13.

Signed-off-by: Chuanhong Guo <gch981213@gmail.com>

* waypoint_follower: use get_parameter without default value

the two-parameter version of get_parameter constructs a
Parameter from the &parameter references. Our value is
unassigned, which caused a warning.
fix 'may be used uninitialized' warning on gcc13.

Signed-off-by: Chuanhong Guo <gch981213@gmail.com>

* nav2_behavior_tree: use get_parameter without default value

the two-parameter version of get_parameter constructs a
Parameter from the &parameter references. Our value is
unassigned, which caused a warning.
fix 'may be used uninitialized' warning on gcc13.

Signed-off-by: Chuanhong Guo <gch981213@gmail.com>

* constrained_smoother: initialize prelast_dir in upsampleAndPopulate

It's technically not needed but GCC can't figure this out and
complains.

Signed-off-by: Chuanhong Guo <gch981213@gmail.com>

* waypoint_follower: fix warns about null pointer dereferencing

gcc13 warns that get_current_goal() may be nullptr and doesn't
allow the assignment of it to poses:
error: potential null pointer dereference [-Werror=null-dereference]

Check the pointer before using it.

Signed-off-by: Chuanhong Guo <gch981213@gmail.com>

* Update nav2_waypoint_follower/src/waypoint_follower.cpp

Signed-off-by: Steve Macenski <stevenmacenski@gmail.com>

---------

Signed-off-by: Chuanhong Guo <gch981213@gmail.com>
Signed-off-by: Steve Macenski <stevenmacenski@gmail.com>
Co-authored-by: Steve Macenski <stevenmacenski@gmail.com>